- Billy.. help!!! Your compiler stinks! (compiler fixes related to namespacing and xml)

+++ source/shared_lib/include/xml/xml_parser.h
@@ -101,8 +101,8 @@ public:
 	~XmlNode();
 
 	const string &getName() const	{return name;}
-	int getChildCount() const		{return children.size();}
-	int getAttributeCount() const	{return attributes.size();}
+	size_t getChildCount() const		{return children.size();}
+	size_t getAttributeCount() const	{return attributes.size();}
 	const string &getText() const	{return text;}
 
 	XmlAttribute *getAttribute(int i) const;
